re-arrange layout
